Create useCreateRoom custom hook
